# specs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Customer Service Demo API
  version: 1.0.0
  description: >
    Frozen V1 API draft for the Web-first e-commerce AI customer service demo.
    All endpoints are scoped to an anonymous Demo Workspace.
servers:
  - url: /api
security:
  - DemoWorkspaceToken: []
...
paths:
...
  /shops/{shopId}:
    parameters:
      - $ref: "#/components/parameters/ShopId"
    get:
      tags: [Shops]
      summary: Get a shop
      responses:
        "200":
          description: Shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    patch:
      tags: [Shops]
      summary: Update mutable shop fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
...
  /shops/{shopId}/products:
    parameters:
      - $ref: "#/components/parameters/ShopId"
    get:
      tags: [Products]
      summary: List products for shop
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /conversations:
    get:
      tags: [Conversations]
      summary: List conversations
      parameters:
        - in: query
          name: shopId
          schema:
            type: string
        - in: query
          name: state
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationSummary"
  /conversations/{conversationId}:
    parameters:
      - $ref: "#/components/parameters/ConversationId"
    get:
      tags: [Conversations]
      summary: Get complete conversation snapshot
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationSnapshot"
...
  /buyer/messages:
    post:
      tags: [Buyer Simulator]
      summary: Send a buyer message into MockDouyinAdapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyerMessageInput"
      responses:
        "202":
          description: Message accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
  /buyer/messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/MessageId"
    patch:
      tags: [Buyer Simulator]
      summary: Edit a buyer message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
      responses:
        "202":
          description: Edit accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
  /buyer/messages/{messageId}/recall:
    parameters:
      - $ref: "#/components/parameters/MessageId"
    post:
      tags: [Buyer Simulator]
      summary: Recall a buyer message
      responses:
        "202":
          description: Recall accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
  /buyer/cards/product:
    post:
      tags: [Buyer Simulator]
      summary: Send a product card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shopId, buyerId, productId]
              properties:
                shopId:
                  type: string
                buyerId:
                  type: string
                productId:
                  type: string
      responses:
        "202":
          description: Product card accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
  /buyer/cards/order:
    post:
      tags: [Buyer Simulator]
      summary: Send an order card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shopId, buyerId, orderId]
              properties:
                shopId:
                  type: string
                buyerId:
                  type: string
                orderId:
                  type: string
      responses:
        "202":
          description: Order card accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
...
  /quality/reviews:
    parameters:
      - in: query
        name: conversationId
        required: false
        schema: { type: string }
    get:
      tags: [Quality]
      summary: List manual quality reviews
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QualityReview"
    post:
      tags: [Quality]
      summary: Manually start a conversation quality review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conversationId]
              properties:
                conversationId:
                  type: string
      responses:
        "202":
          description: Review accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
  /quality/reviews/{reviewId}:
    parameters:
      - $ref: "#/components/parameters/ReviewId"
    get:
      tags: [Quality]
      summary: Get one quality review snapshot
      responses:
        "200":
          description: Quality review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityReview"
        "404":
          $ref: "#/components/responses/NotFound"
  /quality/reviews/{reviewId}/conclusion:
    parameters:
      - $ref: "#/components/parameters/ReviewId"
    post:
      tags: [Quality]
      summary: Submit the human conclusion for an actionable quality review
      description: The review resource determines workspace, tenant and shop scope; callers submit only the bounded result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QualityConclusionInput"
      responses:
        "202":
          description: Human conclusion operation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    DemoWorkspaceToken:
      type: apiKey
      in: header
      name: X-Demo-Workspace-Token
  responses:
    NotFound:
      description: Resource not found in the current Demo Workspace
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  parameters:
    ShopId:
      in: path
      name: shopId
      required: true
      schema: { type: string }
...
    ConversationId:
      in: path
      name: conversationId
      required: true
      schema: { type: string }
...
    MessageId:
      in: path
      name: messageId
      required: true
      schema: { type: string }
...
    ReviewId:
      in: path
      name: reviewId
      required: true
      schema: { type: string }
...
  schemas:
...
    ShopAIMode:
      type: string
      enum: [AUTO_ALLOWED, ASSIST_ONLY, MANUAL_ONLY]
    ConversationMode:
      type: string
      enum: [AUTO, ASSIST, MANUAL, HOLD]
    Shop:
      type: object
      required: [id, name, platform, aiMode, connectionState]
      properties:
        id: { type: string }
        name: { type: string }
        platform: { type: string, enum: [DOUYIN_DEMO] }
        aiMode:
          $ref: "#/components/schemas/ShopAIMode"
        connectionState:
          type: string
          enum: [CONNECTED, RECONNECTING, RECONCILING, DEGRADED, DISCONNECTED]
        enabled: { type: boolean }
        todayConversationCount: { type: integer }
...
    Product:
      type: object
      required: [id, shopId, externalProductId, title, status]
      properties:
        id: { type: string }
        shopId: { type: string }
        externalProductId: { type: string }
        title: { type: string }
        description: { type: string }
        status: { type: string }
        recommendable: { type: boolean }
        learningStatus: { type: string }
        skus:
          type: array
          items:
            $ref: "#/components/schemas/ProductSku"
    ProductSku:
      type: object
      required: [id, externalSkuId, price, inventory]
      properties:
        id: { type: string }
        externalSkuId: { type: string }
        attributes:
          type: object
          additionalProperties:
            type: string
        price: { type: number }
        inventory: { type: integer }
        status: { type: string }
...
    ReplyJobStatus:
      type: string
      enum: [PENDING, GENERATING, FAST_PATH_READY, WAITING_HUMAN, SENT, CANCELLING, STALE, EXPIRED, CANCELLED, FAILED, RECOVERY_PENDING]
    ReplyDraftStatus:
      type: string
      enum: [GENERATING, WAITING_HUMAN, STALE, EXPIRED, FAILED, SENT, CANCELLED]
    ReplyDraftEditType:
      type: string
      enum: [STYLE_EDIT, FACTUAL_CORRECTION, KNOWLEDGE_ENRICHMENT]
    ReplyDraft:
      type: object
      required: [id, replyJobId, aiDraft, status, sourceContextVersion]
      properties:
        id: { type: string }
        replyJobId: { type: string }
        aiDraft: { type: string }
        humanFinal: { type: string, nullable: true }
        editType:
          allOf:
            - $ref: "#/components/schemas/ReplyDraftEditType"
          nullable: true
        status:
          $ref: "#/components/schemas/ReplyDraftStatus"
        sourceContextVersion: { type: integer }
        sourceLastMessageId: { type: string, nullable: true }
        sourceSequence: { type: integer, nullable: true }
        generatedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
        staleReason: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
    ReplyJob:
      type: object
      required: [id, conversationId, status, mode]
      properties:
        id: { type: string }
        conversationId: { type: string }
        userTurnId: { type: string, nullable: true }
        status:
          $ref: "#/components/schemas/ReplyJobStatus"
        mode:
          $ref: "#/components/schemas/ConversationMode"
        sourceLastMessageId: { type: string, nullable: true }
        sourceSequence: { type: integer, nullable: true }
        sourceContextVersion: { type: integer, nullable: true }
        needsReplanReason: { type: string, nullable: true }
        abortReason: { type: string, nullable: true }
        provider: { type: string, nullable: true }
        model: { type: string, nullable: true }
        promptVersion: { type: string, nullable: true }
        ragStrategy: { type: string, nullable: true }
        tokenUsage:
          type: object
          properties:
            inputTokens: { type: integer }
            outputTokens: { type: integer }
        fallbackUsed: { type: boolean }
        draft:
          allOf:
            - $ref: "#/components/schemas/ReplyDraft"
          nullable: true
        currentDraft:
          allOf:
            - $ref: "#/components/schemas/ReplyDraft"
          nullable: true
        sendOutbox:
          allOf:
            - $ref: "#/components/schemas/SendOutbox"
          nullable: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SendOutboxStatus:
      type: string
      enum: [PENDING, SENDING, SENT, FAILED, UNCERTAIN, CANCELLED]
    SendOutbox:
      type: object
      required: [id, idempotencyKey, status]
      properties:
        id: { type: string }
        conversationId: { type: string }
        replyJobId: { type: string, nullable: true }
        idempotencyKey: { type: string }
        expectedLastMessageId: { type: string, nullable: true }
        expectedSequence: { type: integer, nullable: true }
        expectedContextVersion: { type: integer, nullable: true }
        status:
          $ref: "#/components/schemas/SendOutboxStatus"
        receipt:
          type: object
          nullable: true
          properties:
            externalMessageId: { type: string }
            platformMessageId: { type: string }
            sentAt: { type: string, format: date-time }
        failureCode: { type: string, nullable: true }
        failureReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Task:
      type: object
      required: [id, userTurnId, intent, riskLevel, requiredContext, status, blocking]
      properties:
        id: { type: string }
        userTurnId: { type: string }
        intent: { type: string }
        riskLevel: { type: string, enum: [LOW, MEDIUM, HIGH] }
        requiredContext:
          type: array
          items: { type: string }
        requiredKnowledge:
          type: array
          items: { type: string }
        requiredTools:
          type: array
          items: { type: string }
        ownerWorkflowRunId: { type: string, nullable: true }
        status: { type: string }
        result: { type: object, nullable: true }
        blocking: { type: boolean }
    TaskBundle:
      type: object
      required: [id, userTurnId, status, tasks]
      properties:
        id: { type: string }
        userTurnId: { type: string }
        status: { type: string, enum: [ALL_RESOLVED, PARTIAL_RESOLVED, NEEDS_CLARIFICATION, HIGH_RISK, FAILED] }
        tasks:
          type: array
          maxItems: 4
          items:
            $ref: "#/components/schemas/Task"
...
    ConversationSummary:
      type: object
      required: [id, shopId, buyer, state, effectiveMode, contextVersion]
      properties:
        id: { type: string }
        shopId: { type: string }
        buyer:
          $ref: "#/components/schemas/Buyer"
        state: { type: string }
        effectiveMode:
          $ref: "#/components/schemas/ConversationMode"
        contextVersion: { type: integer }
        humanActive: { type: boolean }
        needsReplan: { type: boolean }
        activeReplyJob:
          $ref: "#/components/schemas/ReplyJob"
        sendOutbox:
          $ref: "#/components/schemas/SendOutbox"
        unreadCount: { type: integer }
        lastMessage:
          $ref: "#/components/schemas/Message"
    ConversationSnapshot:
      allOf:
        - $ref: "#/components/schemas/ConversationSummary"
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
            productContext:
              type: object
            customerContext:
              type: object
            summary:
              type: object
            currentDraft:
              $ref: "#/components/schemas/ReplyDraft"
            taskBundle:
              $ref: "#/components/schemas/TaskBundle"
            customerMemories:
              type: array
              items:
                $ref: "#/components/schemas/CustomerMemory"
    Message:
      type: object
      required: [id, conversationId, role, kind, status, sequence, sentAt]
      properties:
        id: { type: string }
        conversationId: { type: string }
        externalMessageId: { type: string }
        sequence: { type: integer }
        role: { type: string, enum: [BUYER, AI, HUMAN, SYSTEM] }
        kind: { type: string, enum: [TEXT, IMAGE, GOODS_CARD, ORDER_CARD, SYSTEM] }
        status: { type: string, enum: [ACTIVE, RECALLED, DELETED, EDITED] }
        content:
          type: object
        sentAt: { type: string, format: date-time }
        receivedAt: { type: string, format: date-time }
    Buyer:
      type: object
      required: [id, externalBuyerId, displayName]
      properties:
        id: { type: string }
        externalBuyerId: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string }
        tags:
          type: array
          items: { type: string }
    CustomerMemoryInput:
      type: object
      required: [shopId, type, key, value]
      properties:
        shopId: { type: string }
        type:
          type: string
          enum: [PREFERENCE, PRODUCT_PREFERENCE, ONGOING_CASE]
        key: { type: string }
        value:
          type: object
        expiresAt:
          type: string
          format: date-time
    CustomerMemory:
      allOf:
        - $ref: "#/components/schemas/CustomerMemoryInput"
        - type: object
          required: [id, buyerId, status]
          properties:
            id: { type: string }
            buyerId: { type: string }
            status: { type: string, enum: [ACTIVE, DISABLED, DELETED] }
...
    BuyerMessageInput:
      type: object
      required: [shopId, buyerId, kind]
      properties:
        shopId: { type: string }
        buyerId: { type: string }
        kind: { type: string, enum: [TEXT, IMAGE] }
        text: { type: string }
        attachmentId: { type: string }
        sentAt: { type: string, format: date-time }
        forcedSequence: { type: integer }
        duplicateExternalMessageId: { type: string }
...
    QualityDeterministicCheck:
      type: object
      required: [key, passed]
      properties:
        key: { type: string }
        passed: { type: boolean }
        reason: { type: string }
    QualityDeterministicResult:
      type: object
      required: [passed, checks]
      properties:
        passed: { type: boolean }
        checks:
          type: array
          items:
            $ref: "#/components/schemas/QualityDeterministicCheck"
    QualityJudgeResult:
      type: object
      required: [relevance, completeness, groundedness, tone, risk, result]
      properties:
        relevance: { type: number, minimum: 0, maximum: 1 }
        completeness: { type: number, minimum: 0, maximum: 1 }
        groundedness: { type: number, minimum: 0, maximum: 1 }
        tone: { type: number, minimum: 0, maximum: 1 }
        risk: { type: string, enum: [LOW, MEDIUM, HIGH] }
        result: { type: string, enum: [PASS, FAIL, NEEDS_HUMAN] }
        reasons:
          type: array
          items: { type: string }
    QualityReview:
      type: object
      required: [id, conversationId, status]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        tenantId: { type: string }
        conversationId: { type: string }
        status: { type: string, enum: [PENDING, RUNNING, AUTO_REVIEWED, PASS, FAIL, NEEDS_HUMAN] }
        deterministicResult:
          oneOf:
            - $ref: "#/components/schemas/QualityDeterministicResult"
            - type: 'null'
        judgeResult:
          oneOf:
            - $ref: "#/components/schemas/QualityJudgeResult"
            - type: 'null'
        humanResult: { type: [string, 'null'], enum: [PASS, FAIL, NEEDS_HUMAN, null] }
        sampleSize: { type: integer, minimum: 0 }
        metrics:
          type: object
          additionalProperties: { type: number }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
...
    QualityConclusionInput:
      type: object
      required: [result]
      additionalProperties: false
      properties:
        result: { type: string, enum: [PASS, FAIL, NEEDS_HUMAN] }
...
    OperationAccepted:
      type: object
      required: [operationId, status]
      properties:
        operationId: { type: string }
        status: { type: string, enum: [ACCEPTED, QUEUED] }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object }